Update with image helper
