--- src/Vision/SingleFeatureMethodTrait.php
@@ -1,37 +1,0 @@
-<?php
-/**
- * Copyright 2018 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-namespace Google\Cloud\Vision;
-
-
-trait SingleFeatureMethodTrait
-{
-    private function buildSingleFeatureRequest(
-        $requestClass,
-        $featureClass,
-        $image,
-        $featureType)
-    {
-        $feature = new $featureClass();
-        $feature->setType($featureType);
-        $features = [$feature];
-        $request = new $requestClass();
-        $request->setImage($image);
-        $request->setFeatures($features);
-        return $request;
-    }
-}
